+++ packages/socialshare/src/index.ts
@@ -1,0 +1,23 @@
+import { SOCIAL_SHARE_ATTRIBUTE } from '@global/constants/attributes';
+import { assessScript, initAttributes } from '@global/factory';
+
+import { version } from '../package.json';
+import { init } from './init';
+
+/**
+ * Init
+ */
+initAttributes();
+
+window.fsAttributes[SOCIAL_SHARE_ATTRIBUTE] ||= {};
+
+const { preventsLoad } = assessScript();
+const attribute = window.fsAttributes[SOCIAL_SHARE_ATTRIBUTE];
+
+attribute.version = version;
+
+if (preventsLoad) attribute.init = init;
+else {
+  window.Webflow ||= [];
+  window.Webflow.push(init);
+}